[MIG] product_manufactorer: Migration to 15.0
