+++ product_manufacturer/__manifest__.py
@@ -2,7 +2,7 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 {
     "name": "Product Manufacturer",
-    "version": "14.0.1.0.2",
+    "version": "15.0.1.0.0",
     "summary": "Adds manufacturers and attributes on the product view.",
     "website": "https://github.com/OCA/product-attribute",
     "author": "OpenERP SA, Odoo Community Association (OCA)",
