+++ product_manufacturer/models/__init__.py
@@ -1,3 +1,5 @@
+# Copyright (C) 2004-2009 Tiny SPRL (<http://tiny.be>).
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import product_manufacturer
+from . import product_template
+from . import product_product
