+++ product_manufacturer/models/product_product.py
@@ -1,0 +1,13 @@
+# Copyright (C) 2004-2009 Tiny SPRL (<http://tiny.be>).
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+
+from odoo import fields, models
+
+
+class ProductProduct(models.Model):
+    _inherit = "product.product"
+
+    manufacturer = fields.Many2one(comodel_name="res.partner")
+    manufacturer_pname = fields.Char(string="Manuf. Product Name")
+    manufacturer_pref = fields.Char(string="Manuf. Product Code")
+    manufacturer_purl = fields.Char(string="Manuf. Product URL")
